+++ crates/testing-utils/src/constants.rs
@@ -49,6 +49,8 @@ pub const TEST_PROGRAM_WASM_BYTECODE: &[u8] =
     include_bytes!("../example_barebones_with_auxilary.wasm");
 /// `infinite_loop.wasm` from the `constraints` repo.
 pub const TEST_INFINITE_LOOP_BYTECODE: &[u8] = include_bytes!("../infinite_loop.wasm");
+/// `example_custom_hash.wasm` from the `constraints` repo.
+pub const TEST_PROGRAM_CUSTOM_HASH: &[u8] = include_bytes!("../example_custom_hash.wasm");
 pub const PREIMAGE_SHOULD_SUCCEED: &[u8] = "asdfasdfasdfasdf".as_bytes();
 pub const PREIMAGE_SHOULD_FAIL: &[u8] = "asdf".as_bytes();
 pub const AUXILARY_DATA_SHOULD_SUCCEED: &[u8] = "fdsafdsa".as_bytes();