+++ crates/threshold-signature-server/src/user/tests.rs
@@ -23,6 +23,7 @@ use std::{
 
 use axum::http::StatusCode;
 use bip39::{Language, Mnemonic};
+use ec_runtime::{Runtime, SignatureRequest};
 use entropy_kvdb::{
     clean_tests,
     encrypted_sled::PasswordMethod,
@@ -33,15 +34,17 @@ use entropy_protocol::{
     user::{user_participates_in_dkg_protocol, user_participates_in_signing_protocol},
     KeyParams, PartyId, SessionId, SigningSessionInfo, ValidatorInfo,
 };
-use entropy_shared::{KeyVisibility, OcwMessageDkg};
+use entropy_shared::{HashingAlgorithm, KeyVisibility, OcwMessageDkg};
 use entropy_testing_utils::{
+    chain_api::entropy::runtime_types::bounded_collections::bounded_vec::BoundedVec as OtherBoundedVec,
     constants::{
         ALICE_STASH_ADDRESS, AUXILARY_DATA_SHOULD_FAIL, AUXILARY_DATA_SHOULD_SUCCEED,
         PREIMAGE_SHOULD_FAIL, PREIMAGE_SHOULD_SUCCEED, TEST_INFINITE_LOOP_BYTECODE,
-        TEST_PROGRAM_WASM_BYTECODE, TSS_ACCOUNTS, X25519_PUBLIC_KEYS,
+        TEST_PROGRAM_CUSTOM_HASH, TEST_PROGRAM_WASM_BYTECODE, TSS_ACCOUNTS, X25519_PUBLIC_KEYS,
     },
     substrate_context::{
-        test_context_stationary, test_node_process_testing_state, SubstrateTestingContext,
+        test_context_stationary, test_node_process_testing_state, testing_context,
+        SubstrateTestingContext,
     },
     test_client::update_pointer,
 };
@@ -79,7 +82,10 @@ use x25519_dalek::{PublicKey, StaticSecret};
 
 use super::UserInputPartyInfo;
 use crate::{
-    chain_api::{entropy, get_api, get_rpc, EntropyConfig},
+    chain_api::{
+        entropy, entropy::runtime_types::bounded_collections::bounded_vec::BoundedVec, get_api,
+        get_rpc, EntropyConfig,
+    },
     get_signer,
     helpers::{
         launch::{
@@ -89,10 +95,10 @@ use crate::{
         signing::Hasher,
         substrate::{get_subgroup, return_all_addresses_of_subgroup},
         tests::{
-            check_if_confirmation, create_clients, initialize_test_logger, run_to_block,
-            setup_client, spawn_testing_validators, update_programs,
+            check_if_confirmation, create_clients, initialize_test_logger, remove_program,
+            run_to_block, setup_client, spawn_testing_validators, update_programs,
         },
-        user::send_key,
+        user::{compute_hash, send_key},
     },
     new_user,
     r#unsafe::api::UnsafeQuery,
@@ -128,6 +134,7 @@ async fn test_sign_tx_no_chain() {
         spawn_testing_validators(Some(signing_address.clone()), false).await;
     let substrate_context = test_context_stationary().await;
     let entropy_api = get_api(&substrate_context.node_proc.ws_url).await.unwrap();
+    let rpc = get_rpc(&substrate_context.node_proc.ws_url).await.unwrap();
 
     let program_hash =
         update_programs(&entropy_api, &two.pair(), TEST_PROGRAM_WASM_BYTECODE.to_owned()).await;
@@ -156,6 +163,7 @@ async fn test_sign_tx_no_chain() {
         auxilary_data: Some(hex::encode(AUXILARY_DATA_SHOULD_SUCCEED)),
         validators_info,
         timestamp: SystemTime::now(),
+        hash: HashingAlgorithm::Keccak,
     };
 
     let validator_ips_and_keys = vec![
@@ -169,10 +177,18 @@ async fn test_sign_tx_no_chain() {
         submit_transaction_requests(validator_ips_and_keys.clone(), generic_msg.clone(), one).await;
 
     for res in test_no_program {
-        assert_eq!(res.unwrap().text().await.unwrap(), "No program set");
+        assert_eq!(res.unwrap().text().await.unwrap(), "No program pointer defined for account");
     }
+    update_pointer(
+        &entropy_api,
+        &rpc,
+        &one.pair(),
+        &one.pair(),
+        OtherBoundedVec(vec![program_hash, program_hash]),
+    )
+    .await
+    .unwrap();
 
-    update_pointer(&entropy_api, &one.pair(), &one.pair(), program_hash).await.unwrap();
     generic_msg.timestamp = SystemTime::now();
     let test_user_res =
         submit_transaction_requests(validator_ips_and_keys.clone(), generic_msg.clone(), one).await;
@@ -185,18 +201,6 @@ async fn test_sign_tx_no_chain() {
         submit_transaction_requests(validator_ips_and_keys.clone(), generic_msg.clone(), one).await;
 
     verify_signature(test_user_res_order, message_hash, keyshare_option.clone()).await;
-
-    generic_msg.timestamp = SystemTime::now();
-    // test failing cases
-    let test_program_pulled =
-        submit_transaction_requests(validator_ips_and_keys.clone(), generic_msg.clone(), two).await;
-
-    for res in test_program_pulled {
-        assert_eq!(
-            res.unwrap().text().await.unwrap(),
-            "Not Registering error: Register Onchain first"
-        );
-    }
 
     generic_msg.timestamp = SystemTime::now();
     let test_user_res_not_registered =
@@ -296,6 +300,20 @@ async fn test_sign_tx_no_chain() {
         );
     }
 
+    // program gets removed and errors
+    remove_program(&entropy_api, &rpc, &two.pair(), program_hash).await;
+    generic_msg.timestamp = SystemTime::now();
+    // test failing cases
+    let test_program_pulled =
+        submit_transaction_requests(validator_ips_and_keys.clone(), generic_msg.clone(), one).await;
+
+    for res in test_program_pulled {
+        assert_eq!(
+            res.unwrap().text().await.unwrap(),
+            format!("No program set at: {program_hash}")
+        );
+    }
+
     let mock_client = reqwest::Client::new();
     // fails verification tests
     // wrong key for wrong validator
@@ -375,9 +393,12 @@ async fn test_fail_signing_group() {
     clean_tests();
 
     let dave = AccountKeyring::Dave;
+    let eve = AccountKeyring::Eve;
     let _ = spawn_testing_validators(None, false).await;
 
-    let _substrate_context = test_node_process_testing_state(false).await;
+    let substrate_context = test_context_stationary().await;
+    let entropy_api = get_api(&substrate_context.node_proc.ws_url).await.unwrap();
+    let rpc = get_rpc(&substrate_context.node_proc.ws_url).await.unwrap();
 
     let validators_info = vec![
         ValidatorInfo {
@@ -393,12 +414,26 @@ async fn test_fail_signing_group() {
         },
     ];
 
+    let program_hash =
+        update_programs(&entropy_api, &eve.pair(), TEST_PROGRAM_WASM_BYTECODE.to_owned()).await;
+    update_pointer(
+        &entropy_api,
+        &rpc,
+        &dave.pair(),
+        &dave.pair(),
+        OtherBoundedVec(vec![program_hash]),
+    )
+    .await
+    .unwrap();
+
     let generic_msg = UserSignatureRequest {
         message: hex::encode(PREIMAGE_SHOULD_SUCCEED),
         auxilary_data: Some(hex::encode(AUXILARY_DATA_SHOULD_SUCCEED)),
         validators_info,
         timestamp: SystemTime::now(),
+        hash: HashingAlgorithm::Keccak,
     };
+
     let server_public_key = PublicKey::from(X25519_PUBLIC_KEYS[0]);
     let signed_message = SignedMessage::new(
         &dave.pair(),
@@ -478,7 +513,7 @@ async fn test_store_share() {
         &alice,
         alice_program.to_account_id().into(),
         KeyVisibility::Public,
-        program_hash,
+        BoundedVec(vec![program_hash]),
     )
     .await;
 
@@ -547,7 +582,7 @@ async fn test_store_share() {
         &alice_program,
         alice_program.to_account_id().into(),
         KeyVisibility::Public,
-        program_hash,
+        BoundedVec(vec![program_hash]),
     )
     .await;
     onchain_user_request.block_number = block_number;
@@ -676,7 +711,7 @@ async fn test_send_and_receive_keys() {
         &alice.clone(),
         alice.to_account_id().into(),
         KeyVisibility::Public,
-        program_hash,
+        BoundedVec(vec![program_hash]),
     )
     .await;
     let block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number;
@@ -733,7 +768,7 @@ pub async fn put_register_request_on_chain(
     sig_req_keyring: &Sr25519Keyring,
     program_modification_account: subxtAccountId32,
     key_visibility: KeyVisibility,
-    program_hash: H256,
+    program_hashes: BoundedVec<H256>,
 ) {
     let sig_req_account =
         PairSigner::<EntropyConfig, sp_core::sr25519::Pair>::new(sig_req_keyring.pair());
@@ -741,7 +776,7 @@ pub async fn put_register_request_on_chain(
     let registering_tx = entropy::tx().relayer().register(
         program_modification_account,
         Static(key_visibility),
-        program_hash,
+        program_hashes,
     );
 
     api.tx()
@@ -770,10 +805,19 @@ async fn test_sign_tx_user_participates() {
         spawn_testing_validators(Some(signing_address.clone()), true).await;
     let substrate_context = test_context_stationary().await;
     let entropy_api = get_api(&substrate_context.node_proc.ws_url).await.unwrap();
+    let rpc = get_rpc(&substrate_context.node_proc.ws_url).await.unwrap();
 
     let program_hash =
         update_programs(&entropy_api, &two.pair(), TEST_PROGRAM_WASM_BYTECODE.to_owned()).await;
-    update_pointer(&entropy_api, &one.pair(), &one.pair(), program_hash).await.unwrap();
+    update_pointer(
+        &entropy_api,
+        &rpc,
+        &one.pair(),
+        &one.pair(),
+        OtherBoundedVec(vec![program_hash]),
+    )
+    .await
+    .unwrap();
 
     let validators_info = vec![
         ValidatorInfo {
@@ -801,6 +845,7 @@ async fn test_sign_tx_user_participates() {
         auxilary_data: Some(hex::encode(AUXILARY_DATA_SHOULD_SUCCEED)),
         validators_info: validators_info.clone(),
         timestamp: SystemTime::now(),
+        hash: HashingAlgorithm::Keccak,
     };
 
     let validator_ips_and_keys = vec![
@@ -1053,7 +1098,7 @@ async fn test_register_with_private_key_visibility() {
         &one,
         program_modification_account.to_account_id().into(),
         KeyVisibility::Private(x25519_public_key),
-        program_hash,
+        BoundedVec(vec![program_hash]),
     )
     .await;
     run_to_block(&rpc, block_number + 1).await;
@@ -1101,6 +1146,34 @@ async fn test_register_with_private_key_visibility() {
 
     assert!(keyshare_result.is_ok());
     clean_tests();
+}
+
+#[tokio::test]
+async fn test_compute_hash() {
+    initialize_test_logger().await;
+    clean_tests();
+    let one = AccountKeyring::Dave;
+    let substrate_context = testing_context().await;
+    let api = get_api(&substrate_context.node_proc.ws_url).await.unwrap();
+    let rpc = get_rpc(&substrate_context.node_proc.ws_url).await.unwrap();
+
+    let mut runtime = Runtime::default();
+    let program_hash =
+        update_programs(&api, &one.pair(), TEST_PROGRAM_CUSTOM_HASH.to_owned()).await;
+
+    let message_hash = compute_hash(
+        &api,
+        &rpc,
+        &HashingAlgorithm::Custom(0),
+        &mut runtime,
+        &[program_hash],
+        PREIMAGE_SHOULD_SUCCEED,
+    )
+    .await
+    .unwrap();
+    // custom hash program uses blake 3 to hash
+    let expected_hash = blake3::hash(PREIMAGE_SHOULD_SUCCEED).as_bytes().to_vec();
+    assert_eq!(message_hash.to_vec(), expected_hash);
 }
 
 pub async fn verify_signature(
@@ -1177,6 +1250,7 @@ async fn test_fail_infinite_program() {
         auxilary_data: Some(hex::encode(AUXILARY_DATA_SHOULD_SUCCEED)),
         validators_info,
         timestamp: SystemTime::now(),
+        hash: HashingAlgorithm::Keccak,
     };
 
     let validator_ips_and_keys = vec![