+++ crates/threshold-signature-server/tests/protocol_wasm.rs
@@ -25,8 +25,9 @@ mod helpers;
 use axum::http::StatusCode;
 use entropy_kvdb::clean_tests;
 use entropy_protocol::{KeyParams, ValidatorInfo};
-use entropy_shared::{KeyVisibility, OcwMessageDkg};
+use entropy_shared::{HashingAlgorithm, KeyVisibility, OcwMessageDkg};
 use entropy_testing_utils::{
+    chain_api::entropy::runtime_types::bounded_collections::bounded_vec::BoundedVec,
     constants::{
         AUXILARY_DATA_SHOULD_SUCCEED, PREIMAGE_SHOULD_SUCCEED, TEST_PROGRAM_WASM_BYTECODE,
         TSS_ACCOUNTS, X25519_PUBLIC_KEYS,
@@ -49,7 +50,6 @@ use std::{
 use subxt::{
     backend::legacy::LegacyRpcMethods,
     ext::sp_core::{sr25519::Signature, Bytes},
-    utils::{AccountId32 as SubxtAccountId32, H256},
     Config, OnlineClient,
 };
 use synedrion::KeyShare;
@@ -74,17 +74,20 @@ async fn test_wasm_sign_tx_user_participates() {
     let one = AccountKeyring::Eve;
     let dave = AccountKeyring::Dave;
 
-    let signing_address = one.clone().to_account_id().to_ss58check();
+    let signing_address = one.to_account_id().to_ss58check();
     let (validator_ips, _validator_ids, users_keyshare_option) =
         spawn_testing_validators(Some(signing_address.clone()), true).await;
     let substrate_context = test_context_stationary().await;
     let entropy_api = get_api(&substrate_context.node_proc.ws_url).await.unwrap();
+    let rpc = get_rpc(&substrate_context.node_proc.ws_url).await.unwrap();
 
     let program_hash =
         update_program(&entropy_api, &dave.pair(), TEST_PROGRAM_WASM_BYTECODE.to_owned())
             .await
             .unwrap();
-    update_pointer(&entropy_api, &one.pair(), &one.pair(), program_hash).await.unwrap();
+    update_pointer(&entropy_api, &rpc, &one.pair(), &one.pair(), BoundedVec(vec![program_hash]))
+        .await
+        .unwrap();
 
     let validators_info = vec![
         ValidatorInfo {
@@ -107,6 +110,7 @@ async fn test_wasm_sign_tx_user_participates() {
         auxilary_data: Some(hex::encode(AUXILARY_DATA_SHOULD_SUCCEED)),
         validators_info: validators_info.clone(),
         timestamp: SystemTime::now(),
+        hash: HashingAlgorithm::Keccak,
     };
 
     let submit_transaction_requests =
@@ -156,7 +160,7 @@ async fn test_wasm_sign_tx_user_participates() {
     .await;
 
     // Check that the signature the user gets matches the first of the server's signatures
-    let user_sig = if let Some(user_sig_stripped) = user_sig.strip_suffix("\n") {
+    let user_sig = if let Some(user_sig_stripped) = user_sig.strip_suffix('\n') {
         user_sig_stripped.to_string()
     } else {
         user_sig
@@ -207,7 +211,7 @@ async fn test_wasm_register_with_private_key_visibility() {
         one.pair(),
         program_modification_account.to_account_id().into(),
         KeyVisibility::Private(x25519_public_key),
-        program_hash,
+        BoundedVec(vec![program_hash]),
     )
     .await
     .unwrap();
@@ -366,7 +370,7 @@ async fn wait_for_register_confirmation(
     account_id: AccountId32,
     api: OnlineClient<EntropyConfig>,
     rpc: LegacyRpcMethods<EntropyConfig>,
-) -> RegisteredInfo<H256, SubxtAccountId32> {
+) -> RegisteredInfo {
     let account_id: <EntropyConfig as Config>::AccountId = account_id.into();
     let registered_query = entropy::storage().relayer().registered(account_id);
     for _ in 0..30 {