+++ crates/threshold-signature-server/tests/sign.rs
@@ -15,6 +15,7 @@
 
 use entropy_kvdb::clean_tests;
 use entropy_testing_utils::{
+    chain_api::entropy::runtime_types::bounded_collections::bounded_vec::BoundedVec,
     constants::{
         AUXILARY_DATA_SHOULD_SUCCEED, PREIMAGE_SHOULD_SUCCEED, TEST_PROGRAM_WASM_BYTECODE,
     },
@@ -39,7 +40,7 @@ async fn integration_test_sign() {
     let pre_registered_user = AccountKeyring::Dave;
     let eve = AccountKeyring::Eve;
 
-    let signing_address = pre_registered_user.clone().to_account_id().to_ss58check();
+    let signing_address = pre_registered_user.to_account_id().to_ss58check();
     let (_validator_ips, _validator_ids, keyshare_option) =
         spawn_testing_validators(Some(signing_address.clone()), false).await;
     let substrate_context = test_context_stationary().await;
@@ -53,9 +54,10 @@ async fn integration_test_sign() {
 
     test_client::update_pointer(
         &api,
+        &rpc,
         &pre_registered_user.pair(),
         &pre_registered_user.pair(),
-        program_hash,
+        BoundedVec(vec![program_hash]),
     )
     .await
     .unwrap();
@@ -89,7 +91,7 @@ async fn integration_test_sign_private() {
     let pre_registered_user = AccountKeyring::Eve;
     let dave = AccountKeyring::Dave;
 
-    let signing_address = pre_registered_user.clone().to_account_id().to_ss58check();
+    let signing_address = pre_registered_user.to_account_id().to_ss58check();
     let (_validator_ips, _validator_ids, keyshare_option) =
         spawn_testing_validators(Some(signing_address.clone()), true).await;
     let substrate_context = test_context_stationary().await;
@@ -103,9 +105,10 @@ async fn integration_test_sign_private() {
 
     test_client::update_pointer(
         &api,
+        &rpc,
         &pre_registered_user.pair(),
         &pre_registered_user.pair(),
-        program_hash,
+        BoundedVec(vec![program_hash]),
     )
     .await
     .unwrap();