+++ pallets/propagation/src/tests.rs
@@ -17,7 +17,7 @@ use std::sync::Arc;
 
 use codec::Encode;
 use entropy_shared::{KeyVisibility, ValidatorInfo};
-use frame_support::{assert_ok, traits::OnInitialize};
+use frame_support::{assert_ok, traits::OnInitialize, BoundedVec};
 use pallet_programs::ProgramInfo;
 use pallet_staking_extension::RefreshInfo;
 use sp_core::offchain::{testing, OffchainDbExt, OffchainWorkerExt, TransactionPoolExt};
@@ -80,17 +80,19 @@ fn knows_how_to_mock_several_http_calls() {
             <Test as frame_system::Config>::Hash::default(),
             ProgramInfo { bytecode: vec![], program_modification_account: 1 },
         );
+        let program_hashes =
+            BoundedVec::try_from(vec![<Test as frame_system::Config>::Hash::default()]).unwrap();
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(1),
             2,
             KeyVisibility::Public,
-            <Test as frame_system::Config>::Hash::default(),
+            program_hashes.clone(),
         ));
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(2),
             3,
             KeyVisibility::Public,
-            <Test as frame_system::Config>::Hash::default(),
+            program_hashes,
         ));
         // full send
         Propagation::post_dkg(4).unwrap();