+++ CHANGELOG.md
@@ -18,9 +18,13 @@ At the moment this project **does not** adhere to
 - When executing the signing protocol on the client-side, a `sig-uid` no longer needs to be given as
   an argument ([#549](https://github.com/entropyxyz/entropy-core/pull/549))
 - Wasm API to entropy-protocol uses camelCase function names ([#566](https://github.com/entropyxyz/entropy-core/pull/566))
+- Wasm API to functions formerly in the x25515chacha20poly1305 repo also now have camelCase function names ([#563](https://github.com/entropyxyz/entropy-core/pull/563))
+- Register and change program pointer interface changed to accept a vecotor of programs. As well pass an index for which containts the hashing code if it custom hashing ([#568](https://github.com/entropyxyz/entropy-core/pull/568))
 
 ### Added
 - Test CLI which calls the same code as in integration tests ([#417](https://github.com/entropyxyz/entropy-core/pull/417))
+- Include contents of x25515chacha20poly1305 repo in entropy-protocol ([#563](https://github.com/entropyxyz/entropy-core/pull/563))
+- Custom Hashing Algorithms [#553](https://github.com/entropyxyz/entropy-core/pull/553/)
 
 ### Changed
 - Crate name refactor ([#561](https://github.com/entropyxyz/entropy-core/pull/561))
@@ -28,6 +32,8 @@ At the moment this project **does not** adhere to
 - Protocol sessions are now identified by a `SessionID` type rather than a `String`
   ([#549](https://github.com/entropyxyz/entropy-core/pull/549))
 - Change bip39 implementation ([#562](https://github.com/entropyxyz/entropy-core/pull/562))
+- Additive programs ([#568](https://github.com/entropyxyz/entropy-core/pull/568))
+- Additional `hash` field in `/sign_tx` JSON body indicates which hashing algorithm to use for signing ([#553](https://github.com/entropyxyz/entropy-core/pull/553))
 
 ### Added